foodgram_backend/requirements.txt
Django==3.2.5
django-filter==2.4.0
djangorestframework==3.12.4
djoser==2.1.0
Pillow==8.3.1
fpdf==1.7.2
gunicorn==20.0.4
psycopg2-binary==2.8.5